Initial commit
Create the basic structure for the library, and add in the structure of the login handshake, as well as some tests to ensure it has the correct size and fields among multiple platforms.

This commit also sets up the automated test workflow, to ensure that commits made to the master branch are properly tested.

Signed-off-by: Triston Plummer <me@triston.io>